prime factor

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,14 +7,14 @@
 
 int main(void)
 {
-	int n, pf;
+	long int n, pf;
 
 	n = 612852475143;
 	for(pf = 2; pf <= n; pf++)
 	{
 		while(n % pf == 0)
 		{
-			printf("%d ", pf);
+			printf("%ld ", pf);
 			n = n / pf;
 		}
 		printf("\n");
